backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
pandas==2.1.0
scikit-learn==1.3.2
pymongo==4.7.0
python-dotenv==1.0.0
certifi==2024.4.0